% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_citations_table.R
\name{create_citation_table}
\alias{create_citation_table}
\title{create citation table}
\usage{
create_citation_table(sheet = curation_sheet())
}
\arguments{
\item{curation_sheet}{a data.frame, the curation worksheet from google sheet (sheet 2 as of now)}
}
\description{
This will be the `citation` data model
}
