% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contrast_table.R
\name{create_contrast_table}
\alias{create_contrast_table}
\title{create contrast table}
\usage{
create_contrast_table(sheet = curation_sheet())
}
\arguments{
\item{sheet}{a data.frame, the curation worksheet from google sheet
(sheet 2 as of now)}
}
\description{
This will be the `contrast` data model
}
