% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diversity_table.R
\name{create_diversity_table}
\alias{create_diversity_table}
\title{create alpha diversity table}
\usage{
create_diversity_table(sheet = curation_sheet())
}
\arguments{
\item{curation_sheet}{a data.frame, the curation worksheet from google sheet (sheet 2 as of now)}
}
\description{
This will be the alpha diversity data model
}
