% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment_table.R
\name{create_experiment_table}
\alias{create_experiment_table}
\title{create experiment table}
\usage{
create_experiment_table(sheet = curation_sheet())
}
\arguments{
\item{curation_sheet}{a data.frame, the curation worksheet from google sheet (sheet 2 as of now)}
}
\description{
This will be the `experiment` data model
}
