% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_keys}
\alias{create_keys}
\title{create_keys}
\usage{
create_keys(prefix, x)
}
\arguments{
\item{prefix}{A prefix}

\item{x}{A list or data.frame}
}
\value{
A key that is the same for any identical rows 
and different for non-identical rows
}
\description{
create_keys
}
\examples{
(df <- data.frame(a=c(1, 1, 5, 6), b=c(1, 1, 5, 9)))
create_keys("key", df)
}
