% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signature_table.R
\name{create_signature_table}
\alias{create_signature_table}
\title{create signature table}
\usage{
create_signature_table(sheet = curation_sheet())
}
\arguments{
\item{curation_sheet}{a data.frame, the curation worksheet from google sheet (sheet 2 as of now)}
}
\description{
This will be the `signature` data model
}
