% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signaturemetadata_table.R
\name{create_signaturemetadata_table}
\alias{create_signaturemetadata_table}
\title{create signature metadata table}
\usage{
create_signaturemetadata_table(sheet = curation_sheet())
}
\arguments{
\item{curation_sheet}{a data.frame, the curation worksheet from google sheet (sheet 2 as of now)}
}
\description{
This will be the `signature metadata` data model
}
