% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statistical_test_table.R
\name{create_statistical_test_table}
\alias{create_statistical_test_table}
\title{create the statistical test table}
\usage{
create_statistical_test_table()
}
\description{
create the statistical test table
}
