% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study_table.R
\name{create_study_table}
\alias{create_study_table}
\title{create study table}
\usage{
create_study_table()
}
\arguments{
\item{sheet}{a data.frame, the curation worksheet from google sheet
(sheet 2 as of now)}
}
\description{
This will be the `study` data model
}
