% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{create_tables}
\alias{create_tables}
\title{create two-table data model}
\usage{
create_tables(sheet = curation_sheet())
}
\arguments{
\item{sheet}{a data.frame, the curation worksheet from google sheet
(sheet 2 as of now)}
}
\description{
This will be the `study` data model
}
