% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheet_manip.R
\name{curation_sheet}
\alias{curation_sheet}
\title{return the curation worksheet (sheet #2)}
\usage{
curation_sheet()
}
\arguments{
\item{sheet}{The spreadsheet reference from \code{\link{google_microbial_sig_sheet}}}
}
\description{
Also skips first row of "data" since that is just a comment
}
