% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologies.R
\name{.read_ontology_from_bioportal_url}
\alias{.read_ontology_from_bioportal_url}
\title{Helper function to do download since files are .gzipped}
\usage{
.read_ontology_from_bioportal_url(bpo_url)
}
\arguments{
\item{bpo_url}{taken from the csv download link of a page like \url{https://bioportal.bioontology.org/ontologies/EFO}}
}
\description{
Helper function to do download since files are .gzipped
}
