% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment_table.R
\name{experimentCols}
\alias{experimentCols}
\title{experimentCols}
\usage{
experimentCols()
}
\value{
A vector with the column names of the experiment table (without key)
}
\description{
experimentCols
}
\examples{
experimentCols()

}
