% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdump.R
\name{get_synonyms}
\alias{get_synonyms}
\title{get and parse synonyms}
\usage{
get_synonyms(download_dir)
}
\arguments{
\item{download_dir}{The directory as prepared by get_synonym_file}
}
\description{
Run after \code{\link{get_synonyms_file}} and supply
the download directory name.
}
