% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdump.R
\name{get_synonyms_file}
\alias{get_synonyms_file}
\title{download taxonomy synonyms}
\usage{
get_synonyms_file(
  download_dir = tempdir(),
 
    taxon_dump_url = "https://data.ace.uq.edu.au/public/gtdb/data/releases/latest/synonyms.tsv"
)
}
\arguments{
\item{download_dir}{do file manipulation here}

\item{taxon_dump_url}{url of `synonyms.tsv` file}
}
\value{
The path to the directory containing the `.tsv` file
}
\description{
download taxonomy synonyms
}
