% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdump.R
\name{get_taxon_dump_files}
\alias{get_taxon_dump_files}
\title{download and untar ncbi taxonomy}
\usage{
get_taxon_dump_files(download_dir = tempdir(),
  taxon_dump_url = "ftp://ftp.ncbi.nlm.nih.gov/pub/taxonomy/taxdump.tar.gz")
}
\arguments{
\item{download_dir}{do file manipulation here}

\item{taxon_dump_url}{url of `taxdump.tar.gz` file at NCBI}
}
\value{
The path to the directory containing the `.dmp`` files
}
\description{
download and untar ncbi taxonomy
}
