% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdump.R
\name{get_taxon_file}
\alias{get_taxon_file}
\title{get taxon file}
\usage{
get_taxon_file(download_dir, taxon_file = "names")
}
\arguments{
\item{download_dir}{The directory as prepared by get_taxon_dump_files}

\item{taxon_file}{The string name (without tht .dmp) of the file to parse, such as 'name'}
}
\description{
Generic parser for weird taxon dmp format
}
\details{
Run after \code{\link{get_taxon_dump_files}} ans supply
the download directory name.
}
