% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxdump.R
\name{get_taxon_names}
\alias{get_taxon_names}
\title{get and parse the `names.dmp` from taxon dump archive}
\usage{
get_taxon_names(download_dir)
}
\arguments{
\item{download_dir}{The directory as prepared by get_taxon_dump_files}
}
\description{
Run after \link{\code{get_taxon_dump_files}} ans supply
the download directory name.
}
