% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metaphlan2ncbi}
\alias{metaphlan2ncbi}
\title{metaphlan2ncbi}
\usage{
metaphlan2ncbi(metaphlan.version = c("2.0", "3.0"))
}
\arguments{
\item{metaphlan.version}{character. Choose between 2.0 or 3.0.}
}
\value{
A named vector storing the mapping from MetaPhlan names to NCBI tax IDs
}
\description{
metaphlan2ncbi
}
