% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study_table.R
\name{studyCols}
\alias{studyCols}
\title{studyCols}
\usage{
studyCols()
}
\value{
A vector with the column names of the study table (without key)
}
\description{
studyCols
}
\examples{
studyCols()

}
