% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheet_manip.R
\name{google_microbial_sig_sheet}
\alias{google_microbial_sig_sheet}
\title{return a reference to the google curation sheet}
\usage{
google_microbial_sig_sheet(key = .ws_key)
}
\arguments{
\item{key}{the google sheet key like in constants.R}
}
\description{
return a reference to the google curation sheet
}
\note{
requires login via `gs_gs()` or will prompt for login.
}
